[Reflection] cleanup ClassTests
