+++ api-source/SomeClass.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 
 
-class SomeClass // extends \Nette\DI\CompilerExtension
+class SomeClass extends \Nette\DI\CompilerExtension
 {
 
 }
