+++ packages/Reflection/tests/Reflection/Class_/ClassReflection/ClassReflectionTest.php
@@ -4,8 +4,10 @@ namespace ApiGen\Reflection\Tests\Reflection\Class_\ClassReflection;
 
 use ApiGen\Reflection\Contract\Reflection\Class_\ClassReflectionInterface;
 use ApiGen\Reflection\Parser\Parser;
+use ApiGen\Reflection\Tests\Reflection\Class_\ClassReflection\Source\ParentClass;
 use ApiGen\Reflection\Tests\Reflection\Class_\ClassReflection\Source\SomeClass;
 use ApiGen\Tests\AbstractParserAwareTestCase;
+use ArrayAccess;
 use phpDocumentor\Reflection\DocBlock\Tags\Author;
 
 final class ClassReflectionTest extends AbstractParserAwareTestCase
@@ -29,8 +31,14 @@ final class ClassReflectionTest extends AbstractParserAwareTestCase
         $this->classReflection = $classReflections[SomeClass::class];
     }
 
+    public function testInterface(): void
+    {
+        $this->assertInstanceOf(ClassReflectionInterface::class, $this->classReflection);
+    }
+
     public function testName(): void
     {
+        $this->assertSame(SomeClass::class, $this->classReflection->getName());
         $this->assertSame('SomeClass', $this->classReflection->getShortName());
         $this->assertSame($this->namespacePrefix, $this->classReflection->getNamespaceName());
     }
@@ -54,5 +62,11 @@ final class ClassReflectionTest extends AbstractParserAwareTestCase
     {
         $this->assertSame(12, $this->classReflection->getStartLine());
         $this->assertSame(23, $this->classReflection->getEndLine());
+    }
+
+    public function testModifiers(): void
+    {
+        $this->assertFalse($this->classReflection->isAbstract());
+        $this->assertFalse($this->classReflection->isFinal());
     }
 }
