+++ packages/ThemeDefault/src/interface.latte
@@ -40,77 +40,77 @@
         </div>
     {/if}
 
-    <div class="info">
-        {var $filteredAnnotations = ($interface->getAnnotations()|annotationFilter)}
-        {foreach $filteredAnnotations as $annotation => $values}
-            {foreach $values as $value}
-                <b>{$annotation}{if $value}:{/if}</b>
-                {$value|annotation:$annotation:$interface|noescape}<br>
-            {/foreach}
-        {/foreach}
-        <a href="{$interface|linkSource}">
-            <strong>Open source code</strong>
-        </a>
-        <br>
-    </div>
+    {*<div class="info">*}
+        {*{var $filteredAnnotations = ($interface->getAnnotations()|annotationFilter)}*}
+        {*{foreach $filteredAnnotations as $annotation => $values}*}
+            {*{foreach $values as $value}*}
+                {*<b>{$annotation}{if $value}:{/if}</b>*}
+                {*{$value|annotation:$annotation:$interface|noescape}<br>*}
+            {*{/foreach}*}
+        {*{/foreach}*}
+        {*<a href="{$interface|linkSource}">*}
+            {*<strong>Open source code</strong>*}
+        {*</a>*}
+        {*<br>*}
+    {*</div>*}
 
-    {if $interface->getOwnMethods()}
-        <table class="summary methods" id="methods">
-            <caption>Methods summary</caption>
-            {foreach $interface->getOwnMethods() as $method}
-                {include "partial/method.latte", method => $method}
-            {/foreach}
-        </table>
-    {/if}
+    {*{if $interface->getOwnMethods()}*}
+        {*<table class="summary methods" id="methods">*}
+            {*<caption>Methods summary</caption>*}
+            {*{foreach $interface->getOwnMethods() as $method}*}
+                {*{include "partial/method.latte", method => $method}*}
+            {*{/foreach}*}
+        {*</table>*}
+    {*{/if}*}
 
-    <table class="summary constants" id="constants" n:if="$interface->getOwnConstants()">
-        <caption>Constants summary</caption>
-        <tr n:foreach="$interface->getOwnConstants() as $constant" id="{$constant->getName()}">
-            {var $annotations = $constant->getAnnotations()}
+    {*<table class="summary constants" id="constants" n:if="$interface->getOwnConstants()">*}
+        {*<caption>Constants summary</caption>*}
+        {*<tr n:foreach="$interface->getOwnConstants() as $constant" id="{$constant->getName()}">*}
+            {*{var $annotations = $constant->getAnnotations()}*}
 
-            <td class="attributes"><code>{$constant->typeHint|typeLinks:$constant|noescape}</code></td>
-            <td class="name">
-                <code>
-                <a href="{$constant|linkSource}"><b>{$constant->getName()}</b></a>
-                </code>
+            {*<td class="attributes"><code>{$constant->typeHint|typeLinks:$constant|noescape}</code></td>*}
+            {*<td class="name">*}
+                {*<code>*}
+                {*<a href="{$constant|linkSource}"><b>{$constant->getName()}</b></a>*}
+                {*</code>*}
 
-                <div class="description detailed">
-                    {$constant->getDescription()}
+                {*<div class="description detailed">*}
+                    {*{$constant->getDescription()}*}
 
-                    {var $filteredAnnotations = ($interface->getAnnotations()|annotationFilter: ['var'])}
-                    {foreach $filteredAnnotations as $annotation => $descriptions}
-                        <h4>{$annotation}</h4>
-                        <div class="list">
-                        {foreach $descriptions as $description}
-                            {if $description}
-                                {$description|annotation:$annotation:$constant|noescape}<br>
-                            {/if}
-                        {/foreach}
-                        </div>
-                    {/foreach}
-                </div>
-            </td>
-            <td class="value">
-                <div>
-                    <a href="#{$constant->getName()}" class="anchor">#</a>
-                    <code>{$constant->valueDefinition|highlightValue:$interface|noescape}</code>
-                </div>
-            </td>
-        </tr>
-    </table>
+                    {*{var $filteredAnnotations = ($interface->getAnnotations()|annotationFilter: ['var'])}*}
+                    {*{foreach $filteredAnnotations as $annotation => $descriptions}*}
+                        {*<h4>{$annotation}</h4>*}
+                        {*<div class="list">*}
+                        {*{foreach $descriptions as $description}*}
+                            {*{if $description}*}
+                                {*{$description|annotation:$annotation:$constant|noescape}<br>*}
+                            {*{/if}*}
+                        {*{/foreach}*}
+                        {*</div>*}
+                    {*{/foreach}*}
+                {*</div>*}
+            {*</td>*}
+            {*<td class="value">*}
+                {*<div>*}
+                    {*<a href="#{$constant->getName()}" class="anchor">#</a>*}
+                    {*<code>{$constant->valueDefinition|highlightValue:$interface|noescape}</code>*}
+                {*</div>*}
+            {*</td>*}
+        {*</tr>*}
+    {*</table>*}
 
-    <table n:foreach="$interface->getInheritedConstants() as $parentName => $inheritedConstants" class="summary inherited">
-        {var $parentClass = ($parentName|getClass)}
-        <caption>Constants inherited from <a href="{$parentName|linkReflection}#constants" n:tag-if="$parentClass">{$parentName}</a></caption>
-        <tr>
-            <td>
-                <code>
-                    {foreach $inheritedConstants as $constant}
-                        <a href="{$constant|constantUrl}" n:tag-if="$parentClass"><b><span n:tag-if="$constant->isDeprecated()" class"deprecated">{$constant->getName()}</span></b></a>{sep}, {/sep}
-                    {/foreach}
-                </code>
-            </td>
-        </tr>
-    </table>
+    {*<table n:foreach="$interface->getInheritedConstants() as $parentName => $inheritedConstants" class="summary inherited">*}
+        {*{var $parentClass = ($parentName|getClass)}*}
+        {*<caption>Constants inherited from <a href="{$parentName|linkReflection}#constants" n:tag-if="$parentClass">{$parentName}</a></caption>*}
+        {*<tr>*}
+            {*<td>*}
+                {*<code>*}
+                    {*{foreach $inheritedConstants as $constant}*}
+                        {*<a href="{$constant|constantUrl}" n:tag-if="$parentClass"><b><span n:tag-if="$constant->isDeprecated()" class"deprecated">{$constant->getName()}</span></b></a>{sep}, {/sep}*}
+                    {*{/foreach}*}
+                {*</code>*}
+            {*</td>*}
+        {*</tr>*}
+    {*</table>*}
 </div>
 {/block}
