+++ packages/ThemeDefault/src/resources/main.js
@@ -1,35 +1,7 @@
-$( function() {
-// $(window).load(function() {
+// $( function() {
+ěě$(window).load(function() {
     var $document = $(document);
     var $groups = $('#groups');
-
-    // Menu
-
-    // Hide deep namespaces
-    $('ul span', $groups).click(function(event) {
-        event.preventDefault();
-        event.stopPropagation();
-        $(this)
-            .toggleClass('collapsed')
-            .parent()
-                .next('ul')
-                    .toggleClass('collapsed');
-    }).click();
-
-    $active = $('ul li.active', $groups);
-    if ($active.length > 0) {
-        // Open active
-        $('> a > span', $active).click();
-    } else {
-        $main = $('> ul > li.main', $groups);
-        if ($main.length > 0) {
-            // Open first level of the main project
-            $('> a > span', $main).click();
-        } else {
-            // Open first level of all
-            $('> ul > li > a > span', $groups).click();
-        }
-    }
 
     // Content
